remove now unnecessary 'dist' folder most files already existed (and were used from) elsewhere. delaunay.js has been moved to external/, and max-links updated to include from there build.py updated to take account of no 'dist' folder (although it does allow for one if it's needed again in the future)

+++ build.py
@@ -117,10 +117,14 @@ outDir = os.path.join('build', buildName)
 if os.path.exists(outDir):
     shutil.rmtree(outDir)
 
-# copy the 'dist' folder - this creates the target directory (and any missing parent dirs)
-# FIXME? replace with manual copy, and any .css and .js files are parsed for replacement tokens?
-shutil.copytree('dist', outDir)
-
+# copy the 'dist' folder, if it exists
+if os.path.exists('dist'):
+    # this creates the target directory (and any missing parent dirs)
+    # FIXME? replace with manual copy, and any .css and .js files are parsed for replacement tokens?
+    shutil.copytree('dist', outDir)
+else:
+    # no 'dist' folder - so create an empty target folder
+    os.makedirs(outDir)
 
 
 # load main.js, parse, and create main total-conversion-build.user.js
